project/requirements.txt
aerich[toml]==0.8.2
asyncpg==0.30.0
fastapi==0.115.12
gunicorn==22.0.0
httpx==0.28.1
pydantic-settings==2.8.1
tortoise-orm==0.25.0
uvicorn==0.34.1
# CV-specific dependencies
tensorflow-cpu==2.17.0
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.20
aiofiles==24.1.0